Feat: 사장 알바 버튼 기능 컴포넌트 분리
